requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
sqlparse==0.3.1
djangorestframework-simplejwt==4.8.0
django-filter==21.1
python-dotenv==0.19.2
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6